# public/swagger.yaml
openapi: 3.0.0
info:
  title: FormBuilder API
  description: "API documentation for FormBuilder application - A form creation and management system. This API uses cookie-based authentication. To authenticate, use the POST /api/auth/login endpoint with your email and password. The server will set an auth_user_id cookie automatically. This cookie will be sent with all subsequent requests automatically. Protected endpoints require this cookie to be present. Note: The authentication cookie is httpOnly, so it cannot be set manually. You must use the login endpoint first. If you receive a 401 Unauthorized error, you need to login using the authentication form at the top of this page or the login endpoint."
  version: 1.0.0
  contact:
    name: FormBuilder API Support
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates a user and sets an authentication cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/forms/{id}:
    get:
      tags:
        - Forms
      summary: Get form by ID
      description: Retrieves a specific form by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the form
      responses:
        '200':
          description: Form details
          content:
            application/json:
              schema:
                type: object
                properties:
                  form:
                    $ref: '#/components/schemas/Form'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Forms
      summary: Update form
      description: Updates an existing form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Updated Form Title
                description:
                  type: string
                  example: Updated description
                fields:
                  type: array
                  items:
                    $ref: '#/components/schemas/FormField'
      responses:
        '200':
          description: Form updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  form:
                    $ref: '#/components/schemas/Form'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Forms
      summary: Delete form
      description: Deletes a form by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the form
      responses:
        '200':
          description: Form deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_user_id
      description: Authentication cookie (auth_user_id) set after login. Use the login endpoint first to authenticate.
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          nullable: true
          example: John Doe
        is_admin:
          type: boolean
          example: false
    Form:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: Customer Feedback Form
        description:
          type: string
          nullable: true
          example: Tell us about your experience
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
        share_token:
          type: string
          example: abc123xyz
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FormField:
      type: object
      properties:
        id:
          type: string
          example: field_1
        type:
          type: string
          enum: [text, email, textarea, radio, checkbox, select, number, date]
          example: text
        label:
          type: string
          example: Full Name
        required:
          type: boolean
          example: true
        options:
          type: array
          items:
            type: string
          description: Options for radio, checkbox, or select fields
          example: ["Excellent", "Good", "Average", "Poor"]
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message here
        success:
          type: boolean
          example: false
        details:
          type: string
          nullable: true
  responses:
    BadRequest:
      description: Bad request - missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'